requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.16
djangorestframework==3.15.2
drf-secure-token==1.2.0
drf-yasg2==1.19.4
idna==3.10
inflection==0.5.1
itypes==1.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pytz==2024.2
requests==2.32.3
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
six==1.16.0
sqlparse==0.5.1
typing-extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
